backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.25",
    "celery[redis]>=5.5.3",
    "django>=5.2.5",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.4",
    "django-filter>=25.1",
    "django-unfold>=0.65.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "itsdangerous>=2.2.0",
    "openfga-sdk>=0.9.5",
    "psycopg2-binary>=2.9.10",
]
